FreeRTOS: Move Files for boardsupport separation and linker change
The build system is making problems with the linker, since it does
not properly override weak symbols. To change this the goal is to
split the plib and add boardsupport objects for the platform specifics

This directory change makes it more clearly visible which parts of the
source are PEARL specific and could be unified with other ports and which
are FreeRTOS specific.

This directory should not contain anything that does not work with FreeRTOS
in general