FreeRTOS: rework the build system of the pear library
The result of this and the previous changes is the separation of PEARL
runtime and the boardsupport. The resulting library and headers need an
additional object File to complete everything necessary to link a
firmware image.

The build system should be somewhat easier to understand,too.

Prepared to build other FreeRTOS ports such as cm3-fpu, cm4, cm0 etc.